[AMD] Remove deprecated Grok performance test and add new accuracy and performance tests
- Deleted the outdated Grok performance benchmark test file.
- Introduced new accuracy tests for DeepSeek-R1, GPT-OSS, GROK, and GSM8K models on MI300X and MI35x.
- Added performance benchmarks for DeepSeek-R1, DeepSeek-V3, and Grok models with various configurations.
- Enhanced CI registration for nightly tests to ensure comprehensive coverage across AMD architectures.
