+++ test/registered/amd/accuracy/test_vlms_mmmu_eval_amd.py
