+++ test/registered/amd/perf/test_deepseek_r1_mxfp4_perf.py
