+++ test/registered/amd/perf/test_deepseek_v31_perf.py
