+++ test/registered/amd/perf/test_deepseek_v3_perf.py
