+++ test/registered/amd/perf/test_grok2_perf.py
@@ -1,0 +1,135 @@
+"""Nightly performance benchmark for Grok-2.
+
+This test benchmarks Grok-2 with FP8 quantization on 8 GPUs.
+
+Model paths can be configured via environment variables:
+- GROK2_MODEL_PATH: Path to Grok-2 model (default: xai-org/grok-2)
+- GROK2_TOKENIZER_PATH: Path to Grok-2 tokenizer (default: alvarobartt/grok-2-tokenizer)
+
+Example usage:
+    python -m pytest test_grok2_perf.py -v
+"""
+
+import os
+import unittest
+from typing import List
+
+from sglang.test.ci.ci_register import register_amd_ci
+from sglang.test.nightly_bench_utils import BenchmarkResult
+from sglang.test.nightly_utils import NightlyBenchmarkRunner
+from sglang.test.test_utils import DEFAULT_URL_FOR_TEST, _parse_int_list_env
+
+# Register for AMD CI - Grok-2 benchmark (~25 min)
+register_amd_ci(est_time=1500, suite="nightly-perf-8-gpu-grok2", nightly=True)
+
+
+def generate_simple_markdown_report(results: List[BenchmarkResult]) -> str:
+    """Generate a simplified markdown report without traces and cost columns."""
+    model_header = results[0].model_path
+    if results[0].run_name and results[0].run_name != "default":
+        model_header += f" ({results[0].run_name})"
+
+    gpu_config = os.getenv("GPU_CONFIG", "")
+    if gpu_config:
+        model_header += f" [{gpu_config}]"
+
+    summary = f"### {model_header}\n"
+    summary += "| batch size | input len | latency (s) | input throughput (tok/s) | output throughput (tok/s) | ITL (ms) |\n"
+    summary += "| ---------- | --------- | ----------- | ------------------------ | ------------------------- | -------- |\n"
+
+    for result in results:
+        itl = 1 / (result.output_throughput / result.batch_size) * 1000
+        summary += f"| {result.batch_size} | {result.input_len} | {result.latency:.2f} | {result.input_throughput:.2f} | {result.output_throughput:.2f} | {itl:.2f} |\n"
+
+    return summary
+
+
+# Model and tokenizer paths can be overridden via environment variables
+GROK2_MODEL_PATH = os.environ.get("GROK2_MODEL_PATH", "xai-org/grok-2")
+GROK2_TOKENIZER_PATH = os.environ.get(
+    "GROK2_TOKENIZER_PATH", "alvarobartt/grok-2-tokenizer"
+)
+PROFILE_DIR = "performance_profiles_grok2"
+
+
+class TestNightlyGrok2Performance(unittest.TestCase):
+    """Nightly performance benchmark for Grok-2.
+
+    Tests Grok-2 with FP8 quantization on TP=8.
+    Runtime: ~25 minutes
+    """
+
+    @classmethod
+    def setUpClass(cls):
+        cls.base_url = DEFAULT_URL_FOR_TEST
+        cls.batch_sizes = [1, 1, 8, 16, 64]
+        cls.input_lens = tuple(_parse_int_list_env("NIGHTLY_INPUT_LENS", "1024"))
+        cls.output_lens = tuple(_parse_int_list_env("NIGHTLY_OUTPUT_LENS", "512"))
+
+        cls.model_config = {
+            "name": "grok2",
+            "model_path": GROK2_MODEL_PATH,
+            "other_args": [
+                "--trust-remote-code",
+                "--tp",
+                "8",
+                "--quantization",
+                "fp8",
+                "--mem-fraction-static",
+                "0.85",
+                "--tokenizer-path",
+                GROK2_TOKENIZER_PATH,
+                "--attention-backend",
+                "aiter",
+            ],
+            "env_vars": {
+                "RCCL_MSCCL_ENABLE": "0",
+                "SGLANG_USE_AITER": "1",
+                "SGLANG_INT4_WEIGHT": "0",
+            },
+        }
+
+        cls.runner = NightlyBenchmarkRunner(PROFILE_DIR, cls.__name__, cls.base_url)
+        cls.runner.setup_profile_directory()
+        cls.runner.full_report = f"## {cls.__name__}\n"
+
+    def test_bench_grok2(self):
+        """Run benchmark for Grok-2."""
+        # Set environment variables
+        old_env = {}
+        for key, value in self.model_config.get("env_vars", {}).items():
+            old_env[key] = os.environ.get(key)
+            os.environ[key] = value
+            print(f"Setting env: {key}={value}")
+
+        try:
+            result_tuple = self.runner.run_benchmark_for_model(
+                model_path=self.model_config["model_path"],
+                batch_sizes=self.batch_sizes,
+                input_lens=self.input_lens,
+                output_lens=self.output_lens,
+                other_args=self.model_config["other_args"],
+                variant=self.model_config["name"],
+                extra_bench_args=["--trust-remote-code"],
+            )
+            results = result_tuple[0]
+            success = result_tuple[1]
+
+            if results:
+                self.runner.full_report += (
+                    generate_simple_markdown_report(results) + "\n"
+                )
+
+            self.assertTrue(success, "Benchmark failed for Grok-2")
+        finally:
+            # Restore original environment
+            for key, value in old_env.items():
+                if value is None:
+                    os.environ.pop(key, None)
+                else:
+                    os.environ[key] = value
+            self.runner.write_final_report()
+
+
+if __name__ == "__main__":
+    unittest.main()
